requirements.txt
langchain
langchain-google-genai
python-dotenv
phonenumbers
email-validator
faiss-cpu
tiktoken
unstructured
pypdf
python-dateutil
streamlit